Add check for not being able to delete the demo map

+++ API/Controllers/MapController.cs
@@ -96,6 +96,11 @@ namespace API.Controllers
         [Route("{id:Guid}")]
         public IActionResult DeleteMap(Guid id)
         {
+            if (id == new Guid("643D4972-F06B-4F4B-5705-08D895F56871"))
+            {
+                return BadRequest("Cannot delete the map for the demo");
+            }
+
             if (_service.DeleteMap(id))
             {
                 return Ok();
